requirements.txt
aiohttp==3.8.1
dnspython==2.2.0
tldextract==3.1.2
requests==2.26.0
